MATLAB image detection project
WINTER 2015 - This project uses gradient histograms that keep track of
bins according to image orientation and direction. Given a training
example image, hog features will be computed and used to train against
test images for what areas on the test image that best match hog
features on the training image.